pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "polymarket-mm"
version = "0.1.0"
description = "Polymarket liquidity rewards market maker bot"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.18.0",
    "aiohttp>=3.9.0",
    "aiohttp-retry>=2.8.3",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.2.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "sortedcontainers>=2.4.0",
    "websockets>=12.0",
    "eth-account>=0.13.0",
    "web3>=6.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.1",
    "freezegun>=1.5.1",
    "ruff>=0.4.7",
    "mypy>=1.10.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
polymarket-mm = "src.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
